added some pr fixes
